requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
Pillow
django-cors-headers
celery
redis
python-decouple
psycopg2-binary
django-filter
requests
